+++ example/DuckDuckGo - advanced matchers/1 - SearchFeature.js
@@ -1,5 +1,5 @@
 description: 'Look up an ambiguous term',
 
 scenario: [
-	SearchBarWidget.searchFor(lookupTerm)
+	SearchBarWidget.searchFor(query)
 ]