+++ example/DuckDuckGo - advanced matchers/2 - ExistenceMatcherFeature.js
@@ -2,6 +2,6 @@ description: 'Make sure the Zero Click Info box exists with booleans',
 
 scenario: [
 	{
-		'ZeroClickWidget.header': true	// prove that the element exists
+		'ZeroClickWidget.meanings': true	// prove that the element exists
 	}
 ]