+++ example/DuckDuckGo - advanced matchers/ZeroClickWidget.js
@@ -1,1 +1,1 @@
-header:	'#zero_click_heading'
+meanings:	'#zci-meanings'