+++ src/controller/Runner.js
@@ -119,7 +119,7 @@ var Runner = new Class( /** @lends Runner# */ {
 	*@private
 	*/
 	initDriver: function initDriver() {
-		if (! this.driver)
+		if (! this.initialized)
 			this.initialized = this.buildDriverFrom(this.config);
 
 		return this.initialized;
@@ -264,32 +264,39 @@ var Runner = new Class( /** @lends Runner# */ {
 		var resolve			= this.deferred.resolve.bind(this.deferred, this),
 			reject			= this.deferred.reject.bind(this.deferred, this.failures),
 			fulfill			= resolve,
-			killDriver		= this.killDriver.bind(this),
+			quitBrowser		= this.quitBrowser.bind(this),
 			precondition	= (this.config.quit == 'always'
-								? killDriver
+								? quitBrowser
 								: promises);	// Q without params simply returns a fulfilled promise
 
 		if (Object.getLength(this.failures) > 0) {
 			fulfill = reject;
 		} else {
 			if (this.config.quit == 'on success')
-				precondition = killDriver;
+				precondition = quitBrowser;
 		}
 
-		precondition().then(fulfill, reject);
+		precondition().done(fulfill, reject);
 	},
 
 	/** Quits the managed browser.
 	*
 	*@return	{QPromise}	A promise resolved once the browser has been properly quit.
 	*/
-	killDriver: function killDriver() {
-		var driver = this.driver;
-		this.driver = null;	// delete reference, as it won't be usable once quitted
+	quitBrowser: function quitBrowser() {
+		var quit = this._quitBrowser.bind(this);
+		return this.initialized ? this.initialized.then(quit, quit) : promises();
+	},
 
-		return (driver	// multiple calls to killDriver() might be issued
-				? this.initialized.then(function() { return driver.quit() })
-				: promises());	// normalize return type to a promise, so that it can safely be called even if the driver had already been quit
+	/** Quits the managed browser immediately, ignoring its availability.
+	*
+	*@return	{QPromise}	A promise resolved once the browser has been quit.
+	*@private
+	*/
+	_quitBrowser: function _quitBrowser() {
+		return this.driver.quit().then(function() {
+			this.initialized = null;
+		}.bind(this));
 	},
 
 	toString: function toString() {