+++ src/view/Runner/Dots.js
@@ -59,7 +59,9 @@ var RunnerDots = new Class({
 							 + getDurationString(this.startTime, new Date())
 							 + ': '
 							 + 'feature'.count(featuresCount)
-							 + ', '
+							 + ' ('
+							 + this.model.config.ignore.length
+							 + ' ignored),'
 							 + 'success'.count(successCount, 'es')
 							 + ', '
 							 + 'failure'.count(failuresCount)
@@ -85,7 +87,7 @@ var getDurationString = function getDurationString(first, second) {
 			second	: Math.floor(durationSeconds) % 60
 		};	// don't you remove this semicolon
 
-	['hour', 'minute', 'second'].forEach(function(unit) {
+	[ 'hour', 'minute', 'second' ].forEach(function(unit) {
 		var value = durations[unit];
 		if (value > 0)
 			result += unit.count(value) + ' ';